deleteComment

+++ public/javascripts/index.js
@@ -28,30 +28,40 @@ window.addEventListener('DOMContentLoaded', () => {
     commentSubmit.forEach((comment) => {
         comment.addEventListener('click', async e => {
             e.preventDefault();
-            const commentList = document.querySelector(`.commentList${e.target.value}`);
+            const postId = e.target.value.split('t')[1];
+            const commentList = document.querySelector(`.commentList${postId}`);
             if(commentList.classList.contains("commentListHidden")) {
                 commentList.classList.remove("commentListHidden")
-                const formField = document.getElementById(`comment${e.target.value}`)
-                const res = await fetch(`/posts/${e.target.value}/comments`, {
+                const formField = document.getElementById(`comment${postId}`)
+                const res = await fetch(`/posts/${postId}/comments`, {
                     method: "POST",
                     headers: {
                         'Content-Type': 'application/json'
                     },
                     body: JSON.stringify({ content: formField.value })
                 })
-                const comments = await res.json()
-    
+                const comments = await res.json();
+                const userId = comments.userId;
+
                     commentList.innerHTML = ""
                     comments.comments.forEach((comment) => {
                         let commentListItem = document.createElement('div');
-                        commentListItem.setAttribute('class', 'commentBox')
-                        commentListItem.innerHTML = comment.content;
+                        let deleteCommentButton = document.createElement('button')
+                        deleteCommentButton.innerHTML = 'Delete';
+                        commentListItem.setAttribute('class', 'commentBox');
+                        deleteCommentButton.setAttribute('class', 'deleteCommentButton');
+                        deleteCommentButton.setAttribute('value', comment.id);
+                        console.log(userId, comment.userId)
                         commentList.appendChild(commentListItem);
-    
+                        commentListItem.innerHTML = comment.content;
+                        if (userId === comment.userId) {
+                            commentListItem.appendChild(deleteCommentButton);
+
+                        }
                     })
 
             }
-            
+
         })
     })
 
@@ -84,13 +94,40 @@ window.addEventListener('DOMContentLoaded', () => {
             }
         })
     //delete a comment
-    document.querySelectorAll('.deleteCommentButton')
+    document.querySelectorAll('.commentsDiv')
         .forEach((button => {
             button.addEventListener("click", async (e) => {
-                const res = await fetch(`/comments/${commentId}`, {
-                    method: "DELETE"
+                if (!e.target.value || e.target.value.startsWith('post')) {
+                    return
+                }
+                const res = await fetch(`/comments/${e.target.value}`, {
+                    method: "DELETE",
+                    headers: {'Content-Type': 'application/json'},
+                    body: JSON.stringify({id: e.target.value})
                 })
                 await res.json();
+                const res = await fetch(`/posts/${e.target.value}/comments`)
+                    const comments = await res.json();
+
+                    const userId = comments.userId;
+
+                    commentList.innerHTML = ""
+                    comments.comments.forEach((comment) => {
+                        let commentListItem = document.createElement('div');
+                        let deleteCommentButton = document.createElement('button')
+                        deleteCommentButton.innerHTML = 'Delete';
+                        commentListItem.setAttribute('class', 'commentBox');
+                        deleteCommentButton.setAttribute('class', 'deleteCommentButton');
+                        deleteCommentButton.setAttribute('value', comment.id);
+                        console.log(userId, comment.userId)
+                        commentList.appendChild(commentListItem);
+                        commentListItem.innerHTML = comment.content;
+                        if (userId === comment.userId) {
+                            commentListItem.appendChild(deleteCommentButton);
+
+                        }
+                    })
+
             })
         }))
 
@@ -103,20 +140,30 @@ window.addEventListener('DOMContentLoaded', () => {
                     commentList.classList.remove("commentListHidden")
                     const res = await fetch(`/posts/${e.target.value}/comments`)
                     const comments = await res.json();
-    
-                        commentList.innerHTML = ""
-                        comments.comments.forEach((comment) => {
-                            let commentListItem = document.createElement('div');
-                            commentListItem.setAttribute('class', 'commentBox')
-                            commentListItem.innerHTML = comment.content;
-                            commentList.appendChild(commentListItem);
-    
-                        })
+
+                    const userId = comments.userId;
+
+                    commentList.innerHTML = ""
+                    comments.comments.forEach((comment) => {
+                        let commentListItem = document.createElement('div');
+                        let deleteCommentButton = document.createElement('button')
+                        deleteCommentButton.innerHTML = 'Delete';
+                        commentListItem.setAttribute('class', 'commentBox');
+                        deleteCommentButton.setAttribute('class', 'deleteCommentButton');
+                        deleteCommentButton.setAttribute('value', comment.id);
+                        console.log(userId, comment.userId)
+                        commentList.appendChild(commentListItem);
+                        commentListItem.innerHTML = comment.content;
+                        if (userId === comment.userId) {
+                            commentListItem.appendChild(deleteCommentButton);
+
+                        }
+                    })
 
                 } else {
                     commentList.classList.add("commentListHidden")
                 }
-                
+
 
             })
         }))
@@ -140,7 +187,7 @@ window.addEventListener('DOMContentLoaded', () => {
                     upvoteDiv.innerHTML = upvotes.upvotes;
                     upvoteDiv.classList.remove("downvoteDiv")
                 }
-    
+
                 })
         })
 })
